[refactor] event publisher, listener 구조 변경
