+++ src/main/java/kr/hhplus/be/server/application/payment/event/publisher/PaymentCreateEventPublisher.java
@@ -1,11 +1,11 @@
-package kr.hhplus.be.server.application.payment.event;
+package kr.hhplus.be.server.application.payment.event.publisher;
 
 import kr.hhplus.be.server.application.payment.dto.PaymentResponse;
 import kr.hhplus.be.server.application.product.dto.ProductResponse;
 
 import java.util.List;
 
-public class PaymentCreateEvent{
+public class PaymentCreateEventPublisher {
 
     public record SendDataPlatform (
             PaymentResponse.Create response
